🔧(renovate) set prCreation to immediate
Previously, prCreation was set to `not-pending` which prevent PR creation until
branch checks succeeded or failed. But for some of our repository, so check jobs
was not run on branch without an associated PR, so renovate never created PR.

Change this option could make renovate noisier but as we have the schedule
option enabled we limit the PR creation frequency to once per week.